requirements.txt
torch>=2.5.1
torchvision>=0.20.1
streamlit>=1.51.0
ultralytics>=8.3.231
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
plotly>=5.15.0
opencv-python>=4.8.0